@@ -4,8 +4,7 @@
       tag="div"
       to="/"
       class="header-abs"
-      v-show="showAbs"
-    >
+      v-show="showAbs">
       <div class="iconfont header-abs-back">&#xe624;</div>
     </router-link>
     <div
@@ -22,71 +21,73 @@
 </template>
 
 <script>
+import { ref, reactive, onMounted, onUnmounted } from 'vue'
 export default {
   name: 'DetailHeader',
-  data () {
-    return {
-      showAbs: true,
-      opacityStyle: {
-        opacity: 0
-      }
-    }
-  },
-  methods: {
-    handleScroll () {
-      const top = document.documentElement.scrollTop
+  setup() {
+    const showAbs = ref(true)
+    const opacityStyle = reactive({
+       opacity: 0
+    })
+
+    function handleScroll () {
+      const top = document.documentElement.scrollTop || document.body.scrollTop || window.pageYOffset
       if (top > 60) {
         let opacity = top / 140
         opacity = opacity > 1 ? 1 : opacity
-        this.opacityStyle = { opacity }
-        this.showAbs = false
+        opacityStyle.opacity = opacity
+        showAbs.value = false
       } else {
-        this.showAbs = true
+        showAbs.value = true
       }
     }
-  },
-  mounted () {
-    window.addEventListener('scroll', this.handleScroll)
-  },
-  beforeDestroy () {
-    window.removeEventListener('scroll', this.handleScroll)
+
+    onMounted(() => {
+      window.addEventListener('scroll', handleScroll)
+    })
+
+    onUnmounted(() => {
+      window.removeEventListener('scroll', handleScroll)
+    })
+
+    return { showAbs, opacityStyle } 
   }
 }
 </script>
 
 <style lang="stylus" scoped>
-@import '~styles/varibles.styl'
-.header-abs
-  position absolute
-  left 0.2rem
-  top 0.2rem
-  width 0.8rem
-  height 0.8rem
-  line-height 0.8rem
-  border-radius 0.4rem
-  text-align center
-  background rgba(0, 0, 0, 0.8)
-  .header-abs-back
-    color #fff
-    font-size 0.4rem
-.header-fixed
-  z-index 2
-  position fixed
-  top 0
-  left 0
-  right 0
-  height $headerHeight
-  line-height $headerHeight
-  text-align center
-  color #fff
-  background $bgColor
-  font-size 0.32rem
-  .header-fixed-back
-    position absolute
-    top 0
-    left 0
-    width 0.64rem
-    text-align center
-    font-size 0.4rem
-    color #fff
+  @import '~styles/varibles.styl'
+  .header-abs
+    position: absolute
+    left: .2rem
+    top: .2rem
+    width: .8rem
+    height: .8rem
+    line-height: .8rem
+    border-radius: .4rem
+    text-align: center
+    background: rgba(0, 0, 0, .8)
+    .header-abs-back
+      color: #fff
+      font-size: .4rem
+  .header-fixed
+    z-index: 2
+    position: fixed
+    top: 0
+    left: 0
+    right: 0
+    height: $headerHeight
+    line-height: $headerHeight
+    text-align: center
+    color: #fff
+    background: $bgColor
+    font-size: .32rem
+    .header-fixed-back
+      position: absolute
+      top: 0
+      left: 0
+      width: .64rem
+      text-align: center
+      font-size: .4rem
+      color: #fff
 </style>
