@@ -9,61 +9,50 @@
 </template>
 
 <script>
+// Composition API
+import HomeHeader from './components/Header'
 import HomeSwiper from './components/Swiper'
 import HomeIcons from './components/Icons'
 import HomeRecommend from './components/Recommend'
 import HomeWeekend from './components/Weekend'
 import axios from 'axios'
-import { mapState } from 'vuex'
+import { useStore } from 'vuex'
+import { ref, onMounted } from 'vue'
 export default {
   name: 'Home',
   components: {
-    // 异步加载组件,项目较小时不建议使用(app.js<1MB)
-    HomeHeader: () => import('./components/Header'),
+    HomeHeader,
     HomeSwiper,
     HomeIcons,
     HomeRecommend,
     HomeWeekend
   },
-  data () {
-    return {
-      lastCity: '',
-      swiperList: [],
-      iconList: [],
-      recommendList: [],
-      weekendList: []
-    }
-  },
-  computed: {
-    ...mapState(['city'])
-  },
-  methods: {
-    getHomeInfo () {
-      axios.get('/api/index.json?city=' + this.city).then(this.getHomeInfoSucc)
-    },
-    getHomeInfoSucc (res) {
+  setup() {
+    const swiperList = ref([])
+    const iconList = ref([])
+    const recommendList = ref([])
+    const weekendList = ref([])
+    const store = useStore()
+    const city = store.state.city
+    async function getHomeInfo() {
+      let res = await axios.get('/api/index.json?city=' + city)
       res = res.data
       if (res.ret && res.data) {
-        const data = res.data
-        this.swiperList = data.swiperList
-        this.iconList = data.iconList
-        this.recommendList = data.recommendList
-        this.weekendList = data.weekendList
+        const result = res.data
+        swiperList.value = result.swiperList
+        iconList.value = result.iconList
+        recommendList.value = result.recommendList
+        weekendList.value = result.weekendList
       }
     }
-  },
-  mounted () {
-    this.lastCity = this.city
-    this.getHomeInfo()
-  },
-  activated () {
-    if (this.lastCity !== this.city) {
-      this.lastCity = this.city
-      this.getHomeInfo()
-    }
+    onMounted(() => {
+      getHomeInfo()
+    })
+    return { swiperList, iconList, recommendList, weekendList }
   }
 }
 </script>
 
 <style>
+
 </style>
