@@ -1,13 +1,7 @@
 <template>
   <div id="app">
     <keep-alive exclude="Detail">
-      <router-view />
+      <router-view/>
     </keep-alive>
   </div>
 </template>
-
-<script>
-export default {
-  name: 'App'
-}
-</script>
