@@ -11,8 +11,8 @@ export default {
 </script>
 
 <style lang="stylus" scoped>
-.v-enter, .v-leave-to
-  opacity 0
-.v-enter-active, .v-leave-active
-  transition opacity 0.5s
+  .v-enter, .v-leave-to
+    opacity: 0
+  .v-enter-active, .v-leave-active
+    transition: opacity .5s
 </style>
