@@ -1,25 +1,27 @@
 <template>
   <div>
     <city-header></city-header>
-    <city-search :cities="cities"></city-search>
+    <city-search :cities="data.cities"></city-search>
     <city-list
-      :cities="cities"
-      :hot="hotCities"
+      :cities="data.cities"
+      :hot="data.hotCities"
       :letter="letter"
     ></city-list>
     <city-alphabet
+      :cities="data.cities"
       @change="handleLetterChange"
-      :cities="cities"
     ></city-alphabet>
   </div>
 </template>
 
 <script>
+import { reactive, ref, onMounted } from 'vue'
 import axios from 'axios'
 import CityHeader from './components/Header'
 import CitySearch from './components/Search'
 import CityList from './components/List'
 import CityAlphabet from './components/Alphabet'
+
 export default {
   name: 'City',
   components: {
@@ -28,42 +30,41 @@ export default {
     CityList,
     CityAlphabet
   },
-  data () {
-    return {
-      cities: {},
-      hotCities: [],
-      letter: ''
-    }
-  },
-  methods: {
-    getCityInfo () {
-      axios.get('/api/city.json').then(this.handleGetCityInfoSucc)
-    },
-    handleGetCityInfoSucc (res) {
-      res = res.data
-      if (res.ret && res.data) {
-        const data = res.data
-        this.cities = data.cities
-        this.hotCities = data.hotCities
-      }
-    },
-    handleLetterChange (letter) {
-      this.letter = letter
-    }
-  },
-  mounted () {
-    this.getCityInfo()
+  setup() {
+    const { letter, handleLetterChange } = useLetterLogic()
+    const { data } = useCityLogic()
+    return { data, letter, handleLetterChange }
   }
 }
+
+function useCityLogic() {
+  const data = reactive({
+    cities: {},
+    hotCities: [],
+  })
+  async function getCityInfo() {
+    let res = await axios.get('/api/city.json')
+    res = res.data
+    if (res.ret && res.data) {
+      const result = res.data
+      data.cities = result.cities
+      data.hotCities = result.hotCities
+    }
+  }
+  onMounted(() => { getCityInfo() })
+  return { data }
+}
+
+function useLetterLogic() {
+  const letter = ref('')
+  function handleLetterChange(selected) {
+    letter.value = selected
+  }
+  return { letter, handleLetterChange }
+}
+
 </script>
 
 <style lang="stylus" scoped>
-@import '~styles/varibles.styl'
-.search
-  height 0.72rem
-  padding 0 0.1rem
-  background $bgColor
-  .search-input
-    height 0.62rem
-    line-height 0.62rem
+
 </style>
