@@ -1,8 +1,5 @@
 <template>
-  <div
-    class="list"
-    ref="wrapper"
-  >
+  <div class="list" ref="wrapper">
     <div>
       <div class="area">
         <div class="title border-topbottom">当前城市</div>
@@ -29,7 +26,7 @@
         class="area"
         v-for="(item, key) of cities"
         :key="key"
-        :ref="key"
+        :ref="elem => elems[key] = elem"
       >
         <div class="title border-topbottom">{{key}}</div>
         <div class="item-list">
@@ -38,7 +35,9 @@
             v-for="innerItem of item"
             :key="innerItem.id"
             @click="handleCityClick(innerItem.name)"
-          >{{innerItem.name}}</div>
+          >
+            {{innerItem.name}}
+          </div>
         </div>
       </div>
     </div>
@@ -46,8 +45,10 @@
 </template>
 
 <script>
+import { watch, onMounted, ref } from 'vue'
 import Bscroll from 'better-scroll'
-import { mapState, mapMutations } from 'vuex'
+import { useStore } from 'vuex'
+import { useRouter } from 'vue-router'
 export default {
   name: 'CityList',
   props: {
@@ -55,69 +56,74 @@ export default {
     cities: Object,
     letter: String
   },
-  computed: {
-    ...mapState({
-      currentCity: 'city'
-    })
-  },
-  methods: {
-    handleCityClick (city) {
-      this.changeCity(city)
-      this.$router.push('/')
-    },
-    ...mapMutations(['changeCity'])
-  },
-  mounted () {
-    this.scroll = new Bscroll(this.$refs.wrapper, { click: true })
-  },
-  watch: {
-    letter () {
-      if (this.letter) {
-        const element = this.$refs[this.letter][0]
-        this.scroll.scrollToElement(element)
-      }
+  setup(props) {
+    const store = useStore()
+    const router = useRouter()
+    const currentCity = store.state.city
+    const elems = ref({})
+    const wrapper = ref(null)
+    const sroll = null
+
+    function handleCityClick(city) {
+      store.commit('changeCity', city)
+      router.push('/')
     }
+
+    watch(() => props.letter, (letter, prevLetter) => {
+      if (letter && scroll) {
+        const element = elems.value[letter]
+        scroll.scrollToElement(element)
+      }
+    })
+
+    onMounted(() => {
+      scroll = new Bscroll(wrapper.value, {
+        click: true
+      })
+    })
+
+    return { elems, wrapper, currentCity, handleCityClick }
   }
 }
 </script>
 
 <style lang="stylus" scoped>
-@import '~styles/varibles.styl'
-.border-topbottom
-  &:before
-    border-color #ccc
-  &:after
-    border-color #ccc
-.border-bottom
-  &:before
-    border-color #ccc
-.list
-  overflow hidden
-  position absolute
-  top 1.58rem
-  left 0
-  right 0
-  bottom 0
-  .title
-    line-height 0.54rem
-    background #eee
-    padding-left 0.2rem
-    color #666
-    font-size 0.26rem
-  .button-list
-    overflow hidden
-    padding 0.1rem 0.6rem 0.1rem 0.1rem
-    .button-wrapper
-      float left
-      width 33.33%
-      .button
-        margin 0.1rem
-        padding 0.1rem 0
-        text-align center
-        border 0.02rem solid #ccc
-        border-radius 0.06rem
-  .item-list
-    .item
-      line-height 0.76rem
-      padding-left 0.2rem
+  @import '~styles/varibles.styl'
+  .border-topbottom
+    &:before
+      border-color: #ccc
+    &:after
+      border-color: #ccc
+  .border-bottom
+    &:before
+      border-color: #ccc
+  .list
+    overflow: hidden
+    position: absolute
+    top: 1.58rem
+    left: 0
+    right: 0
+    bottom: 0
+    .title
+      line-height: .54rem
+      background: #eee
+      padding-left: .2rem
+      color: #666
+      font-size: .26rem
+    .button-list
+      overflow: hidden
+      padding: .1rem .6rem .1rem .1rem
+      .button-wrapper
+        float: left
+        width: 33.33%
+        .button
+          margin: .1rem
+          padding: .1rem 0
+          text-align: center
+          border: .02rem solid #ccc
+          border-radius: .06rem
+    .item-list
+      .item
+        line-height: .76rem
+        padding-left: .2rem
 </style>
