@@ -1,12 +1,7 @@
 <template>
   <div>
     <div class="search">
-      <input
-        class="search-input"
-        type="text"
-        placeholder="输入城市名或拼音"
-        v-model="keyword"
-      />
+      <input v-model="keyword" class="search-input" type="text" placeholder="输入城市名或拼音" />
     </div>
     <div
       class="search-content"
@@ -19,11 +14,10 @@
           v-for="item of list"
           :key="item.id"
           @click="handleCityClick(item.name)"
-        >{{item.name}}</li>
-        <li
-          class="search-item border-bottom"
-          v-show="hasNoData"
         >
+          {{item.name}}
+        </li>
+        <li class="search-item border-bottom" v-show="hasNoData">
           没有找到匹配数据
         </li>
       </ul>
@@ -32,92 +26,90 @@
 </template>
 
 <script>
-import { mapMutations } from 'vuex'
+import { computed, watch, onMounted, ref } from 'vue'
+import { useStore } from 'vuex'
+import { useRouter } from 'vue-router'
 import Bscroll from 'better-scroll'
+
 export default {
   name: 'CitySearch',
   props: {
     cities: Object
   },
-  data () {
-    return {
-      keyword: '',
-      list: [],
-      timer: null
-    }
-  },
-  methods: {
-    handleCityClick (city) {
-      this.changeCity(city)
-      this.$router.push('/')
-    },
-    ...mapMutations(['changeCity'])
-  },
-  computed: {
-    hasNoData () {
-      return !this.list.length
-    }
-  },
-  watch: {
-    keyword () {
-      if (this.timer) {
-        clearTimeout(this.timer)
+  setup(props) {
+    const keyword = ref('')
+    const list = ref([])
+    const search = ref(null)
+    let timer = null
+    const store = useStore()
+    const router = useRouter()
+    const hasNoData = computed(() => {
+      !list.length
+    })
+
+    watch(keyword, (keyword, prevKeyword) => {
+      if (timer) {
+        clearTimeout(timer)
+        timer = null
       }
-      if (!this.keyword) {
-        this.list = []
+      if (!keyword) {
+        list.value = []
         return
       }
-      this.timer = setTimeout(() => {
+      timer = setTimeout(() => {
         const result = []
-        for (let i in this.cities) {
-          this.cities[i].forEach(value => {
-            if (
-              value.spell.indexOf(this.keyword) > -1 ||
-              value.name.indexOf(this.keyword) > -1
-            ) {
+        for (let i in props.cities) {
+          props.cities[i].forEach((value) => {
+            if (value.spell.indexOf(keyword) > -1 || value.name.indexOf(keyword) > -1) {
               result.push(value)
             }
           })
         }
-        this.list = result
+        list.value = result
       }, 100)
-    }
-  },
-  mounted () {
-    this.scroll = new Bscroll(this.$refs.search, {
-      click: true
     })
+
+    function handleCityClick(city) {
+      store.commit('changeCity', city)
+      router.push('/')
+    }
+
+    onMounted(() => {
+      new Bscroll(search.value, { click: true })
+    })
+
+    return { keyword, list, search, hasNoData, handleCityClick }
   }
 }
 </script>
 
 <style lang="stylus" scoped>
-@import '~styles/varibles.styl'
-.search
-  height 0.72rem
-  padding 0 0.1rem
-  background $bgColor
-  .search-input
-    box-sizing border-box
-    width 100%
-    height 0.62rem
-    padding 0 0.1rem
-    line-height 0.62rem
-    text-align center
-    border-radius 0.06rem
-    color #666
-.search-content
-  z-index 1
-  overflow hidden
-  position absolute
-  top 1.58rem
-  left 0
-  right 0
-  bottom 0
-  background #eee
-  .search-item
-    line-height 0.62rem
-    padding-left 0.2rem
-    background #fff
-    color #666
+  @import '~styles/varibles.styl'
+  .search
+    height: .72rem
+    padding: 0 .1rem
+    background: $bgColor
+    .search-input
+      box-sizing: border-box
+      width: 100%
+      height: .62rem
+      padding: 0 .1rem
+      line-height: .62rem
+      text-align: center
+      border-radius: .06rem
+      color: #666
+  .search-content
+    z-index: 1
+    overflow: hidden
+    position: absolute
+    top: 1.58rem
+    left: 0
+    right: 0
+    bottom: 0
+    background: #eee
+    .search-item
+      line-height: .62rem
+      padding-left: .2rem
+      background: #fff
+      color: #666
 </style>
