@@ -9,10 +9,7 @@
         <span class="item-title-icon"></span>
         {{item.title}}
       </div>
-      <div
-        v-if="item.children"
-        class="item-children"
-      >
+      <div v-if="item.children" class="item-chilren">
         <detail-list :list="item.children"></detail-list>
       </div>
     </div>
@@ -22,27 +19,25 @@
 <script>
 export default {
   name: 'DetailList',
-  props: {
-    list: Array
-  }
+  props: { list: Array }
 }
 </script>
 
 <style lang="stylus" scoped>
-.item-title-icon
-  position relative
-  left 0.06rem
-  top 0.06rem
-  display inline-block
-  width 0.36rem
-  height 0.36rem
-  background url('http://s.qunarzz.com/piao/image/touch/sight/detail.png') 0 -0.45rem no-repeat
-  margin-right 0.1rem
-  background-size 0.4rem 3rem
-.item-title
-  line-height 0.8rem
-  font-size 0.32rem
-  padding 0 0.2rem
-.item-children
-  padding 0 0.2rem
+  .item-title-icon
+    position: relative
+    left: .06rem
+    top: .06rem
+    display: inline-block
+    width: .36rem
+    height: .36rem
+    background: url(http://s.qunarzz.com/piao/image/touch/sight/detail.png) 0 -.45rem no-repeat
+    margin-right: .1rem
+    background-size: .4rem 3rem
+  .item-title
+    line-height: .8rem
+    font-size: .32rem
+    padding: 0 .2rem
+  .item-chilren
+    padding: 0 .2rem
 </style>
