@@ -1,17 +1,13 @@
 <template>
   <div>
-    <div
-      class="banner"
-      @click="handleBannerClick"
-    >
-      <img
-        class="banner-img"
-        :src="bannerImg"
-      />
+    <div class="banner" @click="handleBannerClick">
+      <img class="banner-img" :src="bannerImg" />
       <div class="banner-info">
-        <div class="banner-title">{{sightName}}</div>
+        <div class="banner-tittle">
+          {{sightName}}
+        </div>
         <div class="banner-number">
-          <span class="iconfont banner-icon">&#xe695;</span>
+          <span class="iconfont banner-icon">&#xe692;</span>
           {{bannerImgs.length}}
         </div>
       </div>
@@ -25,65 +21,67 @@
     </fade-animation>
   </div>
 </template>
+
 <script>
+import { ref } from 'vue'
 import CommonGallary from 'common/gallary/Gallary'
 import FadeAnimation from 'common/fade/FadeAnimation'
 export default {
   name: 'DetailBanner',
+  components: {
+    CommonGallary,
+    FadeAnimation
+  },
   props: {
     sightName: String,
     bannerImg: String,
     bannerImgs: Array
   },
-  data () {
-    return {
-      showGallary: false
+  setup() {
+    const showGallary = ref(false)
+    
+    function handleBannerClick() {
+      showGallary.value = true
     }
-  },
-  methods: {
-    handleBannerClick () {
-      this.showGallary = true
-    },
-    handleGallaryClose () {
-      this.showGallary = false
+
+    function handleGallaryClose () {
+      showGallary.value = false
     }
-  },
-  components: {
-    CommonGallary,
-    FadeAnimation
+
+    return { showGallary, handleBannerClick, handleGallaryClose }
   }
 }
 </script>
 
 <style lang="stylus" scoped>
-.banner
-  position relative
-  overflow hidden
-  height 0
-  padding-bottom 55%
-  .banner-img
-    width 100%
-  .banner-info
-    display flex
-    position absolute
-    left 0
-    right 0
-    bottom 0
-    line-height 0.6rem
-    color #fff
-    background-image linear-gradient(top, rgba(0, 0, 0, 0), rgba(0, 0, 0, 0.8))
-    .banner-title
-      flex 1
-      font-size 0.32rem
-      padding 0 0.2rem
-    .banner-number
-      height 0.32rem
-      line-height 0.32rem
-      padding 0 0.4rem
-      margin-top 0.14rem
-      border-radius 0.2rem
-      background rgba(0, 0, 0, 0.8)
-      font-size 0.24rem
-      .banner-icon
-        font-size 0.24rem
+  .banner
+    position: relative
+    overflow: hidden
+    height: 0
+    padding-bottom: 55%
+    .banner-img
+      width: 100%
+    .banner-info
+      display: flex
+      position: absolute
+      left: 0
+      right: 0
+      bottom: 0
+      line-height: .6rem
+      color: #fff
+      background-image: linear-gradient(to top, rgba(0, 0, 0, 0), rgba(0, 0, 0, 0.8))
+      .banner-tittle
+        flex: 1
+        font-size: .32rem
+        padding: 0 .2rem
+      .banner-number
+        height: .32rem
+        line-height: .32rem
+        margin-top: .14rem
+        padding: 0 .4rem
+        border-radius: .2rem
+        background: rgba(0, 0, 0, .8)
+        font-size: .24rem
+        .banner-icon
+          font-size: .24rem
 </style>
