@@ -7,7 +7,7 @@
       <span class="iconfont">&#xe632;</span>
       输入城市/景点/游玩主题
     </div>
-    <router-link to="/city">
+    <router-link to='/city'>
       <div class="header-right">
         {{city}}
         <span class="iconfont arrow-icon">&#xe64a;</span>
@@ -17,46 +17,47 @@
 </template>
 
 <script>
-import { mapState } from 'vuex'
+import { useStore } from 'vuex'
 export default {
-  name: 'Header',
-  computed: {
-    ...mapState(['city'])
+  name: 'HomeHeader',
+  setup() {
+    const store = useStore()
+    const city = store.state.city
+    return { city }
   }
 }
 </script>
 
 <style lang="stylus" scoped>
-@import '~styles/varibles.styl'
-/* 1rem = html font-size = 50px 简化rem计算 */
-.header
-  display flex
-  line-height $headerHeight
-  background $bgColor
-  color #fff
-  .header-left
-    width 0.64rem
-    float left
-    .back-icon
-      text-align center
-      font-size 0.4rem
-  .header-input
-    flex 1
-    height 0.64rem
-    line-height 0.64rem
-    margin-top 0.12rem
-    margin-left 0.2rem
-    padding-left 0.2rem
-    background #fff
-    border-radius 0.1rem
-    color #ccc
-  .header-right
-    min-width 1.04rem
-    padding 0 0.1rem
-    float right
-    text-align center
-    color #fff
-    .arrow-icon
-      margin-left -0.04rem
-      font-size 0.24rem
+  @import '~styles/varibles.styl'
+  .header
+    display: flex
+    line-height: $headerHeight
+    background: $bgColor
+    color: #fff
+    .header-left
+      width: .64rem
+      float: left
+      .back-icon
+        text-align: center
+        font-size: .4rem
+    .header-input
+      flex: 1
+      height: .64rem
+      line-height: .64rem
+      margin-top: .12rem
+      margin-left: .2rem
+      padding-left: .2rem
+      background: #fff
+      border-radius: .1rem
+      color: #ccc
+    .header-right
+      min-width: 1.04rem
+      padding: 0 .1rem
+      float: right
+      text-align: center
+      color: #fff
+      .arrow-icon
+        margin-left: -.04rem
+        font-size: .24rem
 </style>
