@@ -9,10 +9,7 @@
         :key="item.id"
         :to="'/detail/' + item.id"
       >
-        <img
-          class="item-img"
-          :src="item.imgUrl"
-        />
+        <img class="item-img" :src="item.imgUrl" />
         <div class="item-info">
           <p class="item-title">{{item.title}}</p>
           <p class="item-desc">{{item.desc}}</p>
@@ -26,44 +23,42 @@
 <script>
 export default {
   name: 'HomeRecommend',
-  props: {
-    list: Array
-  }
+  props: { list: Array }
 }
 </script>
 
 <style lang="stylus" scoped>
-@import '~styles/mixins.styl'
-.title
-  margin-top 0.2rem
-  line-height 0.8rem
-  background #eee
-  text-indent 0.2rem
-.item
-  overflow hidden
-  display flex
-  height 1.9rem
-  .item-img
-    width 1.7rem
-    height 1.7rem
-    padding 0.1rem
-  .item-info
-    flex 1
-    padding 0.1rem
-    min-width 0
-    .item-title
-      line-height 0.54rem
-      font-size 0.32rem
-      ellipsis()
-    .item-desc
-      line-height 0.4rem
-      color #ccc
-      ellipsis()
-    .item-button
-      line-height 0.44rem
-      margin-top 0.16rem
-      background #ff9300
-      padding 0 0.2rem
-      border-radius 0.06rem
-      color #fff
+  @import '~styles/mixins.styl'
+  .title
+    margin-top: .2rem
+    line-height: .8rem
+    background: #eee
+    text-indent: .2rem
+  .item
+    overflow: hidden
+    display: flex
+    height: 1.9rem
+    .item-img
+      width: 1.7rem
+      height: 1.7rem
+      padding: .1rem
+    .item-info
+      flex: 1
+      padding: .1rem
+      min-width: 0
+      .item-title
+        line-height: .54rem
+        font-size: .32rem
+        ellipsis()
+      .item-desc
+        line-height: .4rem
+        color: #ccc
+        ellipsis()
+      .item-button
+        line-height: .44rem
+        margin-top: .16rem
+        background: #ff9300
+        padding: 0 .2rem
+        border-radius: .06rem
+        color: #fff
 </style>
