@@ -8,10 +8,7 @@
         :key="item.id"
       >
         <div class="item-img-wrapper">
-          <img
-            class="item-img"
-            :src="item.imgUrl"
-          />
+          <img class="item-img" :src="item.imgUrl" />
         </div>
         <div class="item-info">
           <p class="item-title">{{item.title}}</p>
@@ -25,32 +22,30 @@
 <script>
 export default {
   name: 'HomeWeekend',
-  props: {
-    list: Array
-  }
+  props: { list: Array }
 }
 </script>
 
 <style lang="stylus" scoped>
-@import '~styles/mixins.styl'
-.title
-  line-height 0.8rem
-  background #eee
-  text-indent 0.2rem
-.item-img-wrapper
-  overflow hidden
-  height 0
-  padding-bottom 37.09%
-  .item-img
-    width 100%
-.item-info
-  padding 0.1rem
-  .item-title
-    line-height 0.54rem
-    font-size 0.32rem
-    ellipsis()
-  .item-desc
-    line-height 0.4rem
-    color #ccc
-    ellipsis()
+  @import '~styles/mixins.styl'
+  .title
+    line-height: .8rem
+    background: #eee
+    text-indent: .2rem
+  .item-img-wrapper
+    overflow: hidden
+    height: 0
+    padding-bottom: 37.09%
+    .item-img
+      width: 100%
+  .item-info
+    padding: .1rem
+    .item-title
+      line-height: .54rem
+      font-size: .32rem
+      ellipsis()
+    .item-desc
+      line-height: .4rem
+      color: #ccc
+      ellipsis()
 </style>
