@@ -3,7 +3,7 @@ try {
   if (localStorage.city) {
     defaultCity = localStorage.city
   }
-} catch (e) { }
+} catch (e) {}
 
 export default {
   city: defaultCity
