@@ -1,7 +1,6 @@
-const path = require('path')
+const path = require('path');
 
 module.exports = {
-  publicPath: process.env.NODE_ENV === 'production' ? '/travel' : '/',
   devServer: {
     proxy: {
       '/api': {
@@ -15,6 +14,7 @@ module.exports = {
   chainWebpack: (config) => {
     config.resolve.alias
       .set('styles', path.join(__dirname, './src/assets/styles/'))
+      .set('@', path.join(__dirname, './src/'))
       .set('common', path.join(__dirname, './src/common/'))
   }
 }
